ノート追加・編集画面からの削除の前に、確認ダイアログを表示する

+++ lib/screen/note_add_edit/note_add_edit_page.dart
@@ -71,10 +71,37 @@ class NoteAddEditPage extends HookWidget {
                   }
                 },
                 deleteCallback: () async {
-                  if (note != null) {
-                    await provider.deleteNote(note.id);
-                  }
-                  Navigator.pop(context); // TODO ここではonWillPopは呼ばれない　らしい(調べる)
+                  // 削除確認ダイアログ表示
+                  showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return AlertDialog(
+                          title: Text('このノートを削除しますか？'),
+                          content: Text('この操作は取り消せません。'),
+                          actions: [
+                            // ボタン領域
+                            FlatButton(
+                              child: Text(
+                                "キャンセル",
+                              ),
+                              onPressed: () async => Navigator.pop(context),
+                            ),
+                            FlatButton(
+                              child: Text(
+                                "削除",
+                                style: TextStyle(color: Colors.red),
+                              ),
+                              onPressed: () async {
+                                if (note != null) {
+                                  await provider.deleteNote(note.id);
+                                }
+                                Navigator.pop(context);
+                                Navigator.pop(context);
+                              },
+                            ),
+                          ],
+                        );
+                      });
                 },
               )
             ],
